+++ wemake_python_styleguide/violations/best_practices.py
@@ -92,6 +92,7 @@ Summary
    WrongEmptyLinesCountViolation
    ImportObjectCollisionViolation
    ProblematicFunctionParamsViolation
+   AwaitInLoopViolation
 
 Best practices
 --------------
@@ -172,6 +173,7 @@ Best practices
 .. autoclass:: WrongEmptyLinesCountViolation
 .. autoclass:: ImportObjectCollisionViolation
 .. autoclass:: ProblematicFunctionParamsViolation
+.. autoclass:: AwaitInLoopViolation
 
 """
 
@@ -2968,3 +2970,38 @@ class ProblematicFunctionParamsViolation(ASTViolation):
 
     error_template = 'Found problematic function parameters'
     code = 475
+
+
+@final
+class AwaitInLoopViolation(ASTViolation):
+    """
+    Do not use ``await`` in ``for`` loop.
+
+    Reasoning:
+        There is a better way to control repeated coroutines in ``for`` loops.
+
+    Solution:
+        Use :func:`asyncio.gather`,
+        :func:`asyncio.wait`,
+        or :class:`asyncio.TaskGroup`
+
+    Example::
+
+        # Correct:
+        async def request():
+            tasks = [parse_content(url) for url in urls]
+            parsed_content = await asyncio.gather(*tasks)
+
+        # Wrong:
+        async def request():
+            parsed_content = []
+            for url in urls:
+                result = await parse_content(url)
+                parsed_content.append(result)
+
+    .. versionadded:: 1.1.0
+
+    """
+
+    error_template = 'Found `await` in `for` loop'
+    code = 476